:sparkles: Adds navigation icons
