+++ package.json
@@ -11,6 +11,7 @@
     "leaflet": "^1.7.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-leaflet": "^3.2.2",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
